Successful implementation of P2P! We have implemented a decent prototype of P2P. Huge leaps in what we're able to do with it. So far we have:
Each user has both a server and clients (one client for each connection).
Users can send data to each other. They can send data to multiple users at once. So far tests have been conducted with simple strings AND an entire blockchain. The blockchain was successfully sent and one its blocks hashed, which turned out to have the EXACT SAME hash as the block from where it was sent.
Users will reconnect to each other if the connection was suddenly terminated (buggy as hell).
Some comments for the new blockchain_network class. More documentation will follow for the rest of the classes, but right now I'm tired. :)
Each server will store the latest data from a client in a "communication" list (class attribute). Perhaps there's a better way to this, but for now it's all we have.
Threading works pretty well, but can occasionally cause some quirky behavior in the program. We probably don't have time to fix this.