Final version for hand-in (#2)
* Final iteration before hand-in
Cleaned up the code and refactored it so makes more sense in an architectural sense.
Added more documentation and some test functions.
Methods have been modified to (hopefully) make the code run better.

* New comment lines for tests

* Last few comment and lines I forgot about
Minor comment additions and deleted some unnecessary lines.

* Very last changes before hand-in
Added disclaimer comments to avoid PLAGIARISM.
Changed the data contained in the blocks of the blockchain to be credentials instead of "Bitcoin".
Fixed the bug that made the pending_credentials always be in the genesis block.
Fixed the validate_chain() method, so it actually works as intended now.
Fixed the compare_hashes() method as well. It only compared the hashes of previous blocks instead of the actually blocks it was selected to compare.
Other small changes.